==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

import sys
sys.executable


import os
from src.data import load_raw_data

os.path.isdir('/Users/franciscoteixeirabarbosa/projects/taxi_driver_predictor/taxi_driver_predictor/src'), callable(load_raw_data)




from src.data import load_raw_data

rides = load_raw_data(year=2022)
rides

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
